Several things a work in progress.
Opening the next and previous archives works.

Switch to having the gui always perform final scaling with
ApproxBilinear. It usually only lasts a few hundred milliseconds and
it's better than having a slow GUI.

The manager never stores fast scaled images now. There's a fast path for
the first image in each archive but that's it. The synchronous loading
path will probably be removed except for the resizing case.

Still uncertain if I want upscaling to happen for older archives. It is
probably more expensive than it is worth for most people.