Improve the CPU usage of animated gifs.
This is currently implemented by converting every frame, proactively,
into a Bgra image, which is very memory-inefficient. In the future this
may need to be converted into some kind of custom per-format frame
iteration.

The memory usage is very poor but its not likely to be a major concern
in practice. The memory usage for static images is low enough that gif
memory usage only becomes a concern when multiple large gifs are loaded
into memory. Webms won't have this problem.