Swap in jemalloc in place of glibc.
With jemalloc 5.2 this requires some tuning, so create
.cargo/config.toml.

Removes the need to drop simple vectors in other threads. Dropping a
large Vec<u8> is at least 4 orders of magnitude faster and is now
unlikely to cause missed frames. Allocations can sometimes be reused,
but that wasn't the core focus since it is less likely to be
consistently helpful; unless it saves an allocation for the currently
visible page it doesn't matter.

Peak and persistent memory usage should be largely unaffected since most
long-lived allocations are very large and contiguous. It does take
longer to return unused memory to the OS, with default settings, but still
far less time than Go took, which was acceptable.