Wire up OpenCL for downscaling.
Compared to the portable alternatives it's not as limited as wgpu
(no 32k limit) and just easier to get both fast and working than vulkan.

Wgpu may still be a viable option as images >= 32k in one dimension are
rare, provided I can figure out an equivalent for pinned memory.