Clean up check_pagination.
Allow scrolling to go negative or beyond the bounds of the current page
when more pages are known to exist. This can't fail, and will stop any
hitches from smooth scrolling while it's happening.

If the individual smooth steps are larger than the images hitching will
still happen, but that's too marginal of a case to bother with.