Fix downscaling, yet again.
OpenCL and the CPU should agree modulo rounding differences between GPU
and CPU due to different float accuracy. A difference in a single step
between adjactent tones in a single channel is sometimes expected.

Fix premultiplication in the shader to match the CPU. Avoid clamping
intermediate values which was unintentional and clamp channels after
reversing premultiplication.