Drop textures and videos during idle time.
Textures typically are fast to drop, but can sometimes take 3+ms and, in
rare cases, up to almost 10ms. It should be possible to drop textures in
another thread entirely but it doesn't seem necessary.

Videos need to be dropped in the main Gui thread and can be glacially
slow, often a second or more. Not much to be done there, but using an
idle callback at least makes the Gui appear more responsive.